backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
celery[redis]==5.3.4
redis==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
minio==7.2.0
wechatpy==1.8.18
httpx==0.25.2
qrcode[pil]==7.4.2
python-barcode==0.15.1
Pillow==10.1.0
opencv-python==4.8.1.78
pyzbar==0.1.9
python-docx==1.1.0
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
